run button will render but need to press on each editor view
